Split conservation API classes from wastd/api.py into conservation application.
